app/api/requirements.txt
# FastAPI and web framework
fastapi[standard]==0.112.0
uvicorn[standard]==0.32.0
python-multipart>=0.0.21
sse-starlette==1.8.2

# Data validation and configuration
pydantic==2.11.1
pydantic-settings==2.5.2
python-dotenv>=1.2.1

# LangChain and AI
langchain==1.1.3
langchain-core>=1.2.5
langchain-openai==1.1.3
langchain-deepseek==1.0.1
langgraph>=1.0.5
langgraph-checkpoint>=3.0.1

# Database
aiosqlite==0.20.0
pysqlite3-binary==0.5.4.post2

# HTTP client
httpx==0.28.1

# PDF processing
PyMuPDF==1.24.14

# Local package
common==0.1.0
